Refactor: Change module name
